requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-cors-headers==3.7.0
google-auth==1.27.1
idna==2.10
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.9
pytz==2020.5
requests==2.25.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.2
